[bug] mmc: sdhci-tcc: fix the compilation error
This patch fixes the incompatible pointer type error.

Linked Commit:
ITS:

(cherry picked from commit 3d0ab37ee7a0e572cbc3426271f6d0fda94a84df)

+++ drivers/mmc/host/sdhci-tcc.c
@@ -842,7 +842,7 @@ static int sdhci_tcc_tap_dly_show(struct seq_file *sf, void *data)
 	return 0;
 }
 
-static int sdhci_tcc_tap_dly_store(struct file *file,
+static ssize_t sdhci_tcc_tap_dly_store(struct file *file,
 					const char __user *ubuf, size_t count, loff_t *ppos)
 {
 	struct seq_file *sf = file->private_data;
@@ -882,7 +882,7 @@ static int sdhci_tcc_clk_dly_show(struct seq_file *sf, void *data)
 	return 0;
 }
 
-static int sdhci_tcc_clk_dly_store(struct file *file,
+static ssize_t sdhci_tcc_clk_dly_store(struct file *file,
 					const char __user *ubuf, size_t count, loff_t *ppos)
 {
 	struct seq_file *sf = file->private_data;
